feat: implement dynamic tool schema generation and streaming tool choice functionality
