+++ packages/parser/package.json
@@ -25,7 +25,6 @@
     "@ai-sdk/provider-utils": "3.0.0-alpha.4"
   },
   "devDependencies": {
-    "@types/json-schema": "^7.0.15",
     "@types/node": "^22.15.21",
     "tsup": "^8.5.0"
   }
