+++ packages/parser/src/tool-call-middleware.ts
@@ -1,19 +1,16 @@
-import { JSONSchema7 } from "json-schema";
-
 import type {
   LanguageModelV2Middleware,
-  LanguageModelV2StreamPart,
   LanguageModelV2ToolCall,
   LanguageModelV2Content,
-  LanguageModelV2FunctionTool,
-  LanguageModelV2ProviderDefinedTool,
-  LanguageModelV2,
-  LanguageModelV2Usage,
-  LanguageModelV2FinishReason,
 } from "@ai-sdk/provider";
 import { generateId } from "@ai-sdk/provider-utils";
-import { getPotentialStartIndex, RJSON } from "./utils";
-import { convertToolPrompt } from "./utils/conv-tool-prompt";
+
+import {
+  RJSON,
+  convertToolPrompt,
+  createDynamicIfThenElseSchema,
+} from "./utils";
+import { normalToolStream, toolChoiceStream } from "./stream-handler";
 
 export function createToolMiddleware({
   toolCallTag,
@@ -315,238 +312,5 @@ export function createToolMiddleware({
 
       return baseReturnParams;
     },
-  };
-}
-
-const createDynamicIfThenElseSchema = (
-  tools: (LanguageModelV2FunctionTool | LanguageModelV2ProviderDefinedTool)[]
-): JSONSchema7 => {
-  // Explicitly specify the return type as JSONSchema7
-  let currentSchema: JSONSchema7 = {};
-  const toolNames: string[] = [];
-
-  for (let i = tools.length - 1; i >= 0; i--) {
-    const tool = tools[i];
-
-    if (tool.type === "provider-defined") {
-      throw new Error(
-        "Provider-defined tools are not supported by this middleware. Please use custom tools."
-      );
-    }
-
-    toolNames.unshift(tool.name);
-
-    const toolCondition: JSONSchema7 = {
-      if: {
-        properties: {
-          name: {
-            const: tool.name,
-          },
-        },
-        required: ["name"],
-      },
-      then: {
-        properties: {
-          name: {
-            const: tool.name,
-          },
-          // Cast tool.parameters to JSONSchema7 here.
-          arguments: tool.parameters as JSONSchema7,
-        },
-        required: ["name", "arguments"],
-      },
-    };
-
-    if (Object.keys(currentSchema).length > 0) {
-      toolCondition.else = currentSchema;
-    }
-
-    currentSchema = toolCondition;
-  }
-
-  return {
-    type: "object", // Explicitly specify type as "object"
-    properties: {
-      name: {
-        type: "string",
-        description: "Name of the tool to call",
-        enum: toolNames,
-      },
-      arguments: {
-        type: "object", // By default, arguments is also specified as object type
-        description: "Argument object to be passed to the tool",
-      },
-    },
-    required: ["name", "arguments"],
-    ...currentSchema,
-  };
-};
-
-// TODO: Modify tool calls to be streamed
-async function toolChoiceStream({
-  doGenerate,
-}: {
-  doGenerate: () => ReturnType<LanguageModelV2["doGenerate"]>;
-}) {
-  const result = await doGenerate();
-
-  // Assume result.content[0] contains tool-call information (JSON)
-  const toolJson: any =
-    result.content[0].type === "text" ? JSON.parse(result.content[0].text) : {};
-
-  const toolCallChunk: LanguageModelV2StreamPart = {
-    type: "tool-call",
-    toolCallType: "function",
-    toolCallId: generateId(),
-    toolName: toolJson.name,
-    args: JSON.stringify(toolJson.arguments || {}),
-  };
-
-  const finishChunk: LanguageModelV2StreamPart = {
-    type: "finish",
-    usage: result.usage as LanguageModelV2Usage,
-    finishReason: "tool-calls" as LanguageModelV2FinishReason,
-  };
-
-  const stream = new ReadableStream<LanguageModelV2StreamPart>({
-    start(controller) {
-      controller.enqueue(toolCallChunk);
-      controller.enqueue(finishChunk);
-      controller.close();
-    },
-  });
-
-  return {
-    request: result.request,
-    response: result.response,
-    stream,
-  };
-}
-
-async function normalToolStream({
-  doStream,
-  toolCallTag,
-  toolCallEndTag,
-}: {
-  doStream: () => ReturnType<LanguageModelV2["doStream"]>;
-  toolCallTag: string;
-  toolCallEndTag: string;
-}) {
-  const { stream, ...rest } = await doStream();
-
-  let isFirstToolCall = true;
-  let isFirstText = true;
-  let afterSwitch = false;
-  let isToolCall = false;
-  let buffer = "";
-
-  let toolCallIndex = -1;
-  let toolCallBuffer: string[] = [];
-
-  const transformStream = new TransformStream<
-    LanguageModelV2StreamPart,
-    LanguageModelV2StreamPart
-  >({
-    transform(chunk, controller) {
-      if (chunk.type === "finish") {
-        if (toolCallBuffer.length > 0) {
-          toolCallBuffer.forEach((toolCall) => {
-            try {
-              const parsedToolCall = RJSON.parse(toolCall) as {
-                name: string;
-                arguments: string;
-              };
-
-              controller.enqueue({
-                type: "tool-call",
-                toolCallType: "function",
-                toolCallId: generateId(),
-                toolName: parsedToolCall.name,
-                args: JSON.stringify(parsedToolCall.arguments),
-              });
-            } catch (e) {
-              console.error(`Error parsing tool call: ${toolCall}`, e);
-
-              controller.enqueue({
-                type: "text",
-                text: `Failed to parse tool call: ${e}`,
-              });
-            }
-          });
-        }
-
-        // stop token
-        controller.enqueue(chunk);
-
-        return;
-      } else if (chunk.type !== "text") {
-        controller.enqueue(chunk);
-        return;
-      }
-
-      buffer += chunk.text;
-
-      function publish(text: string) {
-        if (text.length > 0) {
-          const prefix =
-            afterSwitch && (isToolCall ? !isFirstToolCall : !isFirstText)
-              ? "\n" // separator
-              : "";
-
-          if (isToolCall) {
-            if (!toolCallBuffer[toolCallIndex]) {
-              toolCallBuffer[toolCallIndex] = "";
-            }
-
-            toolCallBuffer[toolCallIndex] += text;
-          } else {
-            controller.enqueue({
-              type: "text",
-              text: prefix + text,
-            });
-          }
-
-          afterSwitch = false;
-
-          if (isToolCall) {
-            isFirstToolCall = false;
-          } else {
-            isFirstText = false;
-          }
-        }
-      }
-
-      do {
-        const nextTag = isToolCall ? toolCallEndTag : toolCallTag;
-        const startIndex = getPotentialStartIndex(buffer, nextTag);
-
-        // no opening or closing tag found, publish the buffer
-        if (startIndex == null) {
-          publish(buffer);
-          buffer = "";
-          break;
-        }
-
-        // publish text before the tag
-        publish(buffer.slice(0, startIndex));
-
-        const foundFullMatch = startIndex + nextTag.length <= buffer.length;
-
-        if (foundFullMatch) {
-          buffer = buffer.slice(startIndex + nextTag.length);
-          toolCallIndex++;
-          isToolCall = !isToolCall;
-          afterSwitch = true;
-        } else {
-          buffer = buffer.slice(startIndex);
-          break;
-        }
-      } while (true);
-    },
-  });
-
-  return {
-    stream: stream.pipeThrough(transformStream),
-    ...rest,
   };
 }
