+++ packages/parser/src/utils/index.ts
@@ -1,4 +1,11 @@
+import { convertToolPrompt } from "./conv-tool-prompt";
+import { createDynamicIfThenElseSchema } from "./dynamic-tool-schema";
 import { getPotentialStartIndex } from "./get-potential-start-index";
 import * as RJSON from "./relaxed-json";
 
-export { getPotentialStartIndex, RJSON };
+export {
+  getPotentialStartIndex,
+  convertToolPrompt,
+  createDynamicIfThenElseSchema,
+  RJSON,
+};
